Create Name class file
